cloud cube 32

+++ mwispDBSCAN.py
@@ -889,8 +889,8 @@ class  MWISPDBSCAN(object):
         clusterValue1D = dataCluster[clusterIndex1D]
         Z0, Y0, X0 = clusterIndex1D
 
-        fitsZero = np.zeros_like(dataCluster,dtype=np.float64)
-        fitsZero=fitsZero+noiseMask 
+        fitsZero = np.zeros_like(dataCluster,dtype=np.float32)
+        fitsZero=fitsZero+noiseMask
          # print cloudTB.colnames
         for eachC in cloudTB:
             cloudID = eachC["_idx"]
